fix comments; separate out ginkgo test
